requirements.txt
alembic==1.12.0
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.1.3
fastapi==0.103.2
flower==2.0.1
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
humanize==4.8.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
pika==0.13.0
pluggy==1.3.0
prometheus-client==0.17.1
prometheus-fastapi-instrumentator==6.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.8
zipp==3.17.0
